move static file, update import
